+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `partial-recall-menu` now displays timestamps that are older than 12
   hours by printing a date.
-- `partial-recall-menu` now displays the update count in shades.
+- `partial-recall-menu` now displays the update count using blcok
+  elements.
 
 ### Fixed